updated penilaian ekstrakulikuler dan komentar CK

+++ database/migrations/2024_11_18_105222_create_penilaian_ekskuls_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('penilaian_ekskuls', function (Blueprint $table) {
             $table->id();
-            $table->integer('nilai')->nullable();
+            $table->float('nilai')->nullable();
             $table->foreignId('mapel_id')->constrained('mapels')->onDelete('cascade');
             $table->foreignId('siswa_id')->constrained('siswas')->onDelete('cascade');
             $table->timestamps();
